Add "interactive" mode, that only boots when explicitly told to
In this mode, the Pico is partitioned:
 - Partition A: contains the Pico program (*).
 - Partition B: contains the alternate Pico program (*).
 - Partition DATA: contains the final Minitel ROM to be
   launched.

The Pico program embeds a trampoline ROM (not to be confused
with the one stored in DATA), that will eventually display an
interactive menu. For now, it just contains a simple program
that parks the Minitel CPU until the boot command is received
over client protocol.

* = This schema is known as A/B partitioning, and it will
    enable future support for updating the Pico's own
    firmware over the client protocol.